Fix incorrect cycle counts for some bit-shift ops
On the 68000 and 68010 shift distance affect execution time, so the
cycle cost of shifting must be taken into account. (On the 68020, shift
distance does not affect execution time.)

This commit fixes a bug in the adjustments made to the m68ki_cycles
table, for all shift operations with known shift distance (where the
shift distance is encoded in the instruction word).

Affects cycle counts for ASL/ASR, LSL/LSR, ROL/ROR, ROXL/ROXR on the
68000 and 68010.