Fix MOVEM cycle-counts
The cycle counts in m68k_in.c are representative for the minimum An/Dn
base case, and additional addressing mode cycle costs are added
elsewhere per mode.

But no An/Dn base case is available for MOVEM, so the data were likely
from the "address indirect" mode, causing an additional cost of 4 or 8
cycles for all addressing modes in the "." case in the table.

So we now fill the table with the cycles for a hypothetical register
base case not present for MOVEM, in order to end up with the correct
calculated cycles for the existing addressing modes.