Undo instruction cycles in trap#n exception
As trap#n exceptions are called from within a normal instruction handler
(as opposed to the address error handler that are called via a longjmp),
control will eventually return to the execution loop where the
instructions cycles will be deducted.

This fix adds the cycles for the instructions back to the available
cycles, or we will pay both the cost of the exception, and of the
instruction.