Don't deduct instruction cycles during address errors
The address error handler is called after a longjmp that does not pass
the part of the excecute loop that deducts instruction cycles. This
means we should not try to undo the instruction's cycles here, as other
exception handlers do, because this will incorrectly reduce the cost of
the exception by the number of cycles for the current instruction.

For costly instructions like MUL and DIV, this will in effect result in
a net increase of available cycles.